pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "yostlabs-graphics"
version = "2025.2.25"
description = "3D graphics engine with OpenGL scene graph, rendering primitives, and DearPyGui integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "dearpygui>=2.1.0",
    "freetype-py>=2.5.1",
    "glfw>=2.10.0",
    "PyOpenGL>=3.1.10",
    "yostlabs-math>=2026.2.23",
    "numpy",
]

[project.optional-dependencies]
sensor = [
    "yostlabs>=2026.2.23",
]
dpg = [
    "dearpygui>=2.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["yostlabs"]

[tool.hatch.build.targets.wheel.force-include]
"assets/fonts" = "yostlabs/graphics/fonts"
"assets/models" = "yostlabs/graphics/models"
